Create a `GroupActionHandler` helper to handle group remove and restore UI
